+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [5.16.0] 2025-01-19
+
 - New AI Provider: Agentforce
 - Create Objects AI-powered documentation
   - Summary